shrfinvsys/invsys/requirement.txt
aniso8601==7.0.0
asgiref==3.5.0
astroid==2.9.3
black==22.1.0
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-axes==5.31.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-ipware==4.0.2
django-sendgrid-v5==1.2.0
djangorestframework==3.13.1
filelock==3.4.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pipenv==2022.1.8
platformdirs==2.5.0
promise==2.3
psycopg2==2.9.3
PyJWT==1.7.1
pylint==2.12.2
python-http-client==3.3.6
pytz==2021.3
Rx==1.6.1
sendgrid==6.9.6
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tzdata==2021.5
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
wrapt==1.13.3
